requirements.txt
# Core dependencies
Flask>=3.0.0
Flask-Session>=0.8.0
rdflib>=7.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Production server
gunicorn>=21.0.0

# Testing
pytest>=8.0.0
pytest-flask>=1.3.0
responses>=0.24.0
